Bug 636750 (4/4) - s/float/double/g on few methods/enums/variables. r=bz sr=sicking

+++ content/html/content/src/nsHTMLMediaElement.cpp
@@ -1426,7 +1426,7 @@ PRBool nsHTMLMediaElement::ParseAttribute(PRInt32 aNamespaceID,
        || aAttribute == nsGkAtoms::loopend
        || aAttribute == nsGkAtoms::start
        || aAttribute == nsGkAtoms::end) {
-      return aResult.ParseFloatValue(aValue);
+      return aResult.ParseDoubleValue(aValue);
     }
     else if (ParseImageAttribute(aAttribute, aValue, aResult)) {
       return PR_TRUE;
